+++ lib/api/libhoney.js
@@ -4,9 +4,9 @@ const libhoney = require("libhoney"),
   process = require("process"),
   path = require("path"),
   url = require("url"),
+  crypto = require("crypto"),
   tracker = require("../async_tracker"),
   instrumentation = require("../instrumentation"),
-  { generateSpanId, generateTraceId } = require("../propagation"),
   deterministicSampler = require("../deterministic_sampler"),
   schema = require("../schema"),
   Span = require("./span"),
@@ -18,6 +18,18 @@ const defaultName = "nodejs";
 const incr = (payload, key, val = 1) => {
   payload[key] = (payload[key] || 0) + val;
 };
+
+// id generation
+const SPAN_ID_BYTES = 8;
+const TRACE_ID_BYTES = 16;
+
+function generateTraceId() {
+  return crypto.randomBytes(TRACE_ID_BYTES).toString("hex");
+}
+
+function generateSpanId() {
+  return crypto.randomBytes(SPAN_ID_BYTES).toString("hex");
+}
 
 const getFilteredOptions = opts => {
   let filteredOptions = {};
