+++ lib/propagation.js
@@ -1,6 +1,5 @@
 /* global require, exports, Buffer, __dirname */
 const path = require("path"),
-  crypto = require("crypto"),
   pkg = require(path.join(__dirname, "..", "package.json")),
   debug = require("debug")(`${pkg.name}:propagation`),
   schema = require("./schema");
@@ -11,21 +10,6 @@ exports.REQUEST_ID_HTTP_HEADER = "X-Request-ID";
 
 const VERSION = "1";
 exports.VERSION = VERSION;
-
-// id generation
-exports.generateSpanId = generateSpanId;
-exports.generateTraceId = generateTraceId;
-
-const SPAN_ID_BYTES = 8;
-const TRACE_ID_BYTES = 16;
-
-function generateTraceId() {
-  return crypto.randomBytes(TRACE_ID_BYTES).toString("hex");
-}
-
-function generateSpanId() {
-  return crypto.randomBytes(SPAN_ID_BYTES).toString("hex");
-}
 
 // assumes a header of the form:
